+++ certbot/tests/compat/filesystem_test.py
@@ -16,6 +16,7 @@ except ImportError:
 
 import certbot.tests.util as test_util
 from certbot import lock
+from certbot import util
 from certbot.compat import os
 from certbot.compat import filesystem
 from certbot.tests.util import TempDirTestCase
@@ -306,6 +307,52 @@ class CopyOwnershipTest(test_util.TempDirTestCase):
         mock_chmod.assert_called_once_with(self.probe_path, 0o700)
 
 
+class CheckPermissionsTest(test_util.TempDirTestCase):
+    def setUp(self):
+        super(CheckPermissionsTest, self).setUp()
+        self.probe_path = _create_probe(self.tempdir)
+
+    def test_check_mode(self):
+        self.assertTrue(filesystem.check_mode(self.probe_path, 0o744))
+
+        filesystem.chmod(self.probe_path, 0o700)
+        self.assertFalse(filesystem.check_mode(self.probe_path, 0o744))
+
+    @unittest.skipIf(POSIX_MODE, reason='Test specific to Windows security')
+    def test_check_owner_windows(self):
+        self.assertTrue(filesystem.check_owner(self.probe_path))
+
+        system = win32security.ConvertStringSidToSid(SYSTEM_SID)
+        security = win32security.SECURITY_ATTRIBUTES().SECURITY_DESCRIPTOR
+        security.SetSecurityDescriptorOwner(system, False)
+
+        with mock.patch('win32security.GetFileSecurity') as mock_get:
+            mock_get.return_value = security
+            self.assertFalse(filesystem.check_owner(self.probe_path))
+
+    @unittest.skipUnless(POSIX_MODE, reason='Test specific to Linux security')
+    def test_check_owner_linux(self):
+        self.assertTrue(filesystem.check_owner(self.probe_path))
+
+        import os as std_os  # pylint: disable=os-module-forbidden
+        uid = std_os.getuid()
+
+        with mock.patch('os.getuid') as mock_uid:
+            mock_uid.return_value = uid + 1
+            self.assertFalse(filesystem.check_owner(self.probe_path))
+
+    def test_check_permissions(self):
+        self.assertTrue(filesystem.check_permissions(self.probe_path, 0o744))
+
+        with mock.patch('certbot.compat.filesystem.check_mode') as mock_mode:
+            mock_mode.return_value = False
+            self.assertFalse(filesystem.check_permissions(self.probe_path, 0o744))
+
+        with mock.patch('certbot.compat.filesystem.check_owner') as mock_owner:
+            mock_owner.return_value = False
+            self.assertFalse(filesystem.check_permissions(self.probe_path, 0o744))
+
+
 class OsReplaceTest(test_util.TempDirTestCase):
     """Test to ensure consistent behavior of rename method"""
     def test_os_replace_to_existing_file(self):
@@ -381,7 +428,7 @@ def _set_owner(target, security_owner, user):
 def _create_probe(tempdir):
     filesystem.chmod(tempdir, 0o744)
     probe_path = os.path.join(tempdir, 'probe')
-    open(probe_path, 'w').close()
+    util.safe_open(probe_path, 'w', chmod=0o744).close()
     return probe_path