+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Certbot adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 * CLI flag `--deactivate-authorizations` has been added. It deactivates existing valid authorizations before and after issuing new certificates, ensuring that the results of all challenges are fresh. It is automatically enabled by `--dry-run` and `--staging`, but can be disabled with `--no-deactivate-authorizations`. 
+* Turn off session tickets for apache plugin by default
 * acme: Authz deactivation added to `acme` module.
 
 ### Changed